Fixed weird application of _.partial fn
Seems like the best way to fix the collision issue: if p and q are next
to each other and moving p would cause it to collide with q (the
distance between the centers of p and q is less than or equal to the
sum of the radii of p and q), then don’t move p. Or move p, but then
move it back before re-rendering.